requirements.txt
minecraft-launcher-lib
requests
Pillow
pypresence
git+https://github.com/t-mart/skinpy.git
pystray
